# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse": {
				"properties": {
					"status": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"status",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "learnexo",
		"version": "1.0.0",
		"description": "this is an edtech application for personalized learning using Artificial Intelligence",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "serg lab"
		}
	},
	"paths": {
		"/api/v1/questionnaire/bulk-upload": {
			"post": {
				"operationId": "BulkUploadQuestions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Questionnaire"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/questionnaire/{id}": {
			"put": {
				"operationId": "UpdateQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Questionnaire"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Questionnaire"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/auth/onboarding": {
			"post": {
				"operationId": "Onboarding",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"photo": {
										"type": "string",
										"format": "binary"
									},
									"dateOfBirth": {
										"type": "string"
									},
									"studentClass": {
										"type": "string"
									},
									"gender": {
										"type": "string"
									},
									"town": {
										"type": "string"
									},
									"state": {
										"type": "string"
									},
									"schoolName": {
										"type": "string"
									},
									"schoolAddress": {
										"type": "string"
									},
									"learningStyle": {
										"type": "string"
									},
									"pastExam": {
										"type": "string"
									},
									"language": {
										"type": "string"
									},
									"residentialAddress": {
										"type": "string"
									},
									"stateOfOrigin": {
										"type": "string"
									}
								},
								"required": [
									"photo",
									"dateOfBirth",
									"studentClass",
									"gender",
									"town",
									"state",
									"schoolName",
									"schoolAddress",
									"learningStyle",
									"pastExam",
									"language",
									"residentialAddress",
									"stateOfOrigin"
								]
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "https://dev-api.learnexo.ng",
			"description": "API Gateway development"
		},
		{
			"url": "http://localhost:7070",
			"description": "API Gateway local"
		}
	],
...
}